Move all modules to a module folder
